+++ src/scene/Interactive.h
@@ -126,7 +126,7 @@ long GetNumberInterWithOutScriptLoad();
 
 void UnlinkAllLinkedObjects();
 EntityHandle IsCollidingAnyInter(const Vec3f & pos, const Vec3f & size);
-Entity * GetFirstInterAtPos(const Vec2s & pos, long flag = 0, Vec3f * _pRef = NULL, Entity ** _pTable = NULL, int * _pnNbInTable = NULL);
+Entity * GetFirstInterAtPos(const Vec2s & pos, long flag = 0, Vec3f * _pRef = NULL, Entity ** _pTable = NULL, size_t * _pnNbInTable = NULL);
 
 /*!
  * Adds an Interactive Object to the Scene
